all_eq {unitizer}	R Documentation

Like all.equal but Returns Empty String If Not all.equal

Description

Used as the default value comparison function since when values mismatch we use diffObj which would make the text output from all.equal somewhat redundant.

Usage

all_eq(target, current, ...)

Arguments

target
R object

current
other R object to be compared to target

...
arguments to pass to all.equal

Value

TRUE if all.equal returns TRUE, "" otherwise all_eq(1, 1L) all_eq(1, 2) isTRUE(all_eq(1, 2))

[Package unitizer version 1.4.14 Index]