as.expression,unitizerList-method {unitizer}	R Documentation

Coerce to expression by returning items coerced to expressions

Description

Really only meaningful for classes that implement the .items slot as an expression, but works for others to the extent .items contents are coercible to expressions

Usage

## S4 method for signature 'unitizerList'
as.expression(x, ...)

[Package unitizer version 1.4.14 Index]