capture_output {unitizer}	R Documentation

Capture Both StdOut and StdErr

Description

Will sink both "output" and "message" streams without checking whether they are already sunk, and will unsink them the same way.

Usage

capture_output(expr, env = parent.frame())

## S3 method for class 'captured_output'
print(x, ...)

Arguments

env
an environment to evaluate them in

a
quoted to evaluate

Value

a list with stdout and stderr captured separately, classed as "captured_output"

[Package unitizer version 1.4.14 Index]