nextItem,unitizerList-method {unitizer}	R Documentation

Iterate through items of a unitizerList ObjectJK

Description

Extraction process is a combination of steps:

Move Internal pointer with nextItem or prevItem

Retrieve item getItem

Check whether we're done iterating with done

done will return TRUE if the pointer is on either the first or last entry depending on what direction you are iterating. If you wish to iterate from the last item forward, you should either reset with parameter reverse set to TRUE, or re-order the items.

Usage

## S4 method for signature 'unitizerList'
nextItem(x)

Arguments

x
a unitizerList object

Value

unitizerList for getItem, an item from the list, which could be anything

[Package unitizer version 1.4.14 Index]