print.bullet {unitizer}	R Documentation

Print Methods for UL and OL objects

Description

Print Methods for UL and OL objects

Usage

## S3 method for class 'bullet'
print(x, width = 0L, ...)

Arguments

x
object to print

width
integer how many characters to wrap at, if set to 0 will auto detect width with getOptions("width")

Value

invisibly a character vector with one element per line printed

[Package unitizer version 1.4.14 Index]