repair_environments {unitizer}	R Documentation

Repair Environment Chains

Description

In theory should never be needed, but use in case you get errors about corrupted environments. You should only use this if you get an error telling you to use it.

Usage

repair_environments(x)

Arguments

x
either a unitizer, or a store id (see unitize)

Details

If you pass a store id this will re-save the repaired unitizer to the location specified by the store id.

Value

a unitizer object

See Also

unitize

[Package unitizer version 1.4.14 Index]