run_ls {unitizer}	R Documentation

Worker function to actually execute the 'ls' work

Description

Worker function to actually execute the 'ls' work

Usage

run_ls(env, stop.env, all.names, pattern, store.env = NULL)

Arguments

env
the environment to start `ls`ing in

stop.env
the environment to stop at

all.names,
same as `ls`

pattern
same as `ls`

store.env
NULL or environment, if the latter will populate that environment with all the objects found between `env` and `stop.env`

Value

character or environment depending on `store.env`

[Package unitizer version 1.4.14 Index]