testFuns {unitizer}	R Documentation

Store Functions for New vs. Reference Test Comparisons

Description

testFuns contains the functions used to compare the results and side effects of running test expressions. “testFuns” objects can be used as the compare argument for unitizer_sect, thereby allowing you to specify different comparison functions for different aspects of test evaluation.

Details

The default comparison functions are as follows:

value: all_eq

conditions: all_eq

output: function(x, y) TRUE, i.e. not compared

message: function(x, y) TRUE, i.e. not compared as conditions should be capturing warnings/errors

aborted: function(x, y) TRUE, i.e. not compared as conditions should also be capturing this implicitly

See Also

unitizer_sect for more relevant usage examples, all_eq

Examples

# use `identical` instead of `all.equal` to compare values
testFuns(value=identical)

[Package unitizer version 1.4.14 Index]