unitizer {unitizer}	R Documentation

unitizer

Description

Simplifies regression tests by comparing objects produced by test code with earlier versions of those same objects. If objects are unchanged the tests pass. 'unitizer' provides an interactive interface to review failing tests or new tests. See vignettes for details.

[Package unitizer version 1.4.14 Index]