unitizerChanges-class {unitizer}	R Documentation

Summary of Changes

Description

Changes arise any time a user, through the interactive unitizer mode, requests the storing of a change (accept new version of failed test, add new test, remove old test).

[Package unitizer version 1.4.14 Index]