unitizerSection-class {unitizer}	R Documentation

Contains Representation For a Section of Tests

Description

unitizerSectionExpression-class contains the actual expressions that belong to the section, whereas unitizerSection-class only contains the meta data. The latter objects are used within ]unitizer-class, whereas the former is really just a temporary object until we can generate the latter.

Details

unitizerSectionNA-class is a specialized section for tests that actually don't have a section (removed tests that are nonetheless chosen to be kept by user in interactive environment)

Slots

title

1 lenght character, the name of the section

details

character vector containing additional info on the section

compare

functions to compare the various aspects of a unitizerItem-class @slot length tracks size of the section

[Package unitizer version 1.4.14 Index]