unitizer_ls {unitizer}	R Documentation

An 'ls' Like Function

Description

Much like 'ls', except that it is designed to crawl up the `.new` and `.ref` environments and display all the objects.

Usage

unitizer_ls(
  name,
  pos = -1L,
  envir = parent.frame(),
  all.names = FALSE,
  pattern
)

Details

This is used in `browseUnitizer,unitizer-unitizerBrowse-method`, and is re-assigned to `ls` for use in the `unitizer` prompt environment.

Value

list of object names, or a list with environments containing the objects

[Package unitizer version 1.4.14 Index]