show,unitizerDummy-method {unitizer}	R Documentation

Documentation Block for Internal S4 Methods

Description

R insists these need to be documented as user facing, but they are not really so were throwing them all in here. Actual docs are in non roxygen comments by fun definitions.

Usage

## S4 method for signature 'unitizerDummy'
show(object)

## S4 method for signature 'unitizerGlobalIndices'
as.integer(x, ...)

## S4 method for signature 'unitizerList'
length(x)

## S4 method for signature 'unitizerList,subIndex,missing,missing'
x[i]

## S4 method for signature 'unitizerList,subIndex'
x[[i]]

## S4 replacement method for signature 'unitizerList,subIndex'
x[i] <- value

## S4 replacement method for signature 'unitizerList,subIndex'
x[[i]] <- value

## S4 method for signature 'unitizerList'
as.list(x, ...)

## S4 method for signature 'unitizerList,ANY'
append(x, values, after = length(x))

## S4 method for signature 'unitizerList'
c(x, ..., recursive = FALSE)

## S4 method for signature 'factor,factor'
append(x, values, after = length(x))

## S4 method for signature 'unitizerList'
names(x)

## S4 replacement method for signature 'unitizerList'
names(x) <- value

## S4 method for signature 'unitizerItem'
initialize(.Object, ...)

## S4 method for signature 'unitizerItem'
show(object)

## S4 method for signature 'unitizerItems,unitizerItemOrNULL'
e1 + e2

## S4 method for signature 'unitizerItems,unitizerItems'
e1 + e2

## S4 method for signature 
## 'unitizerTests,unitizerSectionExpressionOrExpression'
e1 + e2

## S4 method for signature 'unitizerSummary'
show(object)

## S4 method for signature 'unitizer'
initialize(.Object, ...)

## S4 method for signature 'unitizer'
length(x)

## S4 method for signature 'unitizer'
summary(object, silent = FALSE, ...)

## S4 method for signature 'unitizerObjectList'
summary(object, silent = FALSE, ...)

## S4 method for signature 'unitizerObjectListSummary'
show(object)

## S4 method for signature 'unitizer'
as.character(x, ...)

## S4 method for signature 'unitizerBrowse'
show(object)

## S4 method for signature 'unitizerBrowse'
as.character(x, width = 0L, ...)

## S4 method for signature 'unitizerBrowse'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

## S4 method for signature 'unitizerBrowse,unitizerBrowseSection'
e1 + e2

## S4 method for signature 'unitizerBrowseSubSection'
length(x)

## S4 method for signature 'unitizerBrowse,subIndex,missing,missing'
x[i]

## S4 method for signature 'unitizerBrowseSubSection,subIndex,missing,missing'
x[i]

## S4 method for signature 'unitizerBrowseSection,unitizerBrowseSubSection'
e1 + e2

## S4 method for signature 'unitizerLoadFail'
show(object)

## S4 method for signature 'unitizerInPkg'
as.character(x, ...)

## S4 method for signature 'unitizerInPkg'
show(object)

## S4 method for signature 'unitizerState'
show(object)

## S4 method for signature 'unitizerDummy,unitizerDummy'
all.equal(target, current, ...)

## S4 method for signature 'unitizerDummy,ANY'
all.equal(target, current, ...)

## S4 method for signature 'ANY,unitizerDummy'
all.equal(target, current, ...)

## S4 method for signature 'unitizerStateRaw,unitizerStateRaw'
all.equal(target, current, ...)

## S4 method for signature 'unitizer,unitizerSection'
e1 + e2

## S4 method for signature 'unitizer,unitizerTestsOrExpression'
e1 + e2

## S4 method for signature 'unitizer,unitizerItems'
e1 + e2

## S4 method for signature 'unitizer,unitizerItem'
e1 + e2

## S4 method for signature 'unitizer,unitizerItemTestsErrors'
e1 + e2

Details

Put in this file because this file is included by almost every other file

[Package unitizer version 1.4.14 Index]